requirements.txt
This project needs installation of the below packages
requests
panda
openpyxl